Fixed issue with 0, 0 Pred
The idea is, setting all previous position pixels to 1, when convoluted, will look like everything is occupied. So, I tested 1 pixel for last position only, which shows that the model preds not 0, 9. I think that if I change the values to 0.1 (any value other than 1) for ALL last pixels rather than 1, it should work